Fix RSC serialization error and client-side server-only imports on Operator Overview page
